+++ packages/main/builds/cdn.js
@@ -1,0 +1,5 @@
+import AlpineI18n from '../src/index'
+
+document.addEventListener('alpine:initializing', () => {
+    AlpineI18n(window.Alpine)
+})