+++ packages/main/builds/module.js
@@ -1,0 +1,3 @@
+import AlpineI18n from '../src/index'
+
+export default AlpineI18n